# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Persona Debate Local API
  version: 1.0.0
  description: |
    Local-first Node.js API for persona/debate/chat/governance workflows.
    All responses use a standard envelope:
    - success: { ok: true, data: ... }
    - error: { ok: false, error: { code, message, details } }
servers:
  - url: http://localhost:3000
security:
  - bearerAuth: []
  - cookieAuth: []
  - apiKeyAuth: []
...
paths:
...
  /api/agentic/router/preview:
    post:
      tags: [Agentic]
      summary: Preview persona team routing
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgenticTeam'
      responses:
        '200': { description: Preview, content: { application/json: { schema: { $ref: '#/components/schemas/OkResponse' } } } }
  /api/agentic/plan:
    post:
      tags: [Agentic]
      summary: Generate task plan from goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgenticPlanRequest'
      responses:
        '200': { description: Planned, content: { application/json: { schema: { $ref: '#/components/schemas/OkResponse' } } } }
  /api/agentic/templates:
    get:
      tags: [Agentic]
      summary: List task templates
      responses:
        '200': { description: Templates, content: { application/json: { schema: { $ref: '#/components/schemas/OkResponse' } } } }
    post:
      tags: [Agentic]
      summary: Save task template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, template]
              properties:
                id: { type: string }
                name: { type: string }
                template:
                  type: object
                  required: [steps]
                  properties:
                    title: { type: string }
                    objective: { type: string }
                    team: { type: object, additionalProperties: true }
                    settings: { type: object, additionalProperties: true }
                    steps:
                      type: array
                      items: { $ref: '#/components/schemas/AgenticTaskStep' }
      responses:
        '201': { description: Saved, content: { application/json: { schema: { $ref: '#/components/schemas/OkResponse' } } } }
  /api/agentic/templates/{templateId}:
    delete:
      tags: [Agentic]
      summary: Delete task template
      parameters:
        - in: path
          name: templateId
          required: true
          schema: { type: string }
      responses:
        '200': { description: Deleted, content: { application/json: { schema: { $ref: '#/components/schemas/OkResponse' } } } }
  /api/agentic/tasks:
    get:
      tags: [Agentic]
      summary: List tasks
      parameters:
        - in: query
          name: status
          schema: { type: string }
      responses:
        '200': { description: Task list, content: { application/json: { schema: { $ref: '#/components/schemas/OkResponse' } } } }
    post:
      tags: [Agentic]
      summary: Create task draft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgenticTaskRequest'
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/OkResponse' } } } }
  /api/agentic/tasks/{taskId}:
    get:
      tags: [Agentic]
      summary: Get task
      parameters:
        - in: path
          name: taskId
          required: true
          schema: { type: string }
      responses:
        '200': { description: Task detail, content: { application/json: { schema: { $ref: '#/components/schemas/OkResponse' } } } }
  /api/agentic/tasks/{taskId}/run:
    post:
      tags: [Agentic]
      summary: Run task
      parameters:
        - in: path
          name: taskId
          required: true
          schema: { type: string }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                maxSteps: { type: integer, minimum: 1, maximum: 500, default: 100 }
      responses:
        '200': { description: Task run result, content: { application/json: { schema: { $ref: '#/components/schemas/OkResponse' } } } }
  /api/agentic/approvals:
    get:
      tags: [Agentic]
      summary: List approvals
      parameters:
        - in: query
          name: status
          schema: { type: string }
      responses:
        '200': { description: Approvals, content: { application/json: { schema: { $ref: '#/components/schemas/OkResponse' } } } }
  /api/agentic/approvals/{approvalId}/decision:
    post:
      tags: [Agentic]
      summary: Approve/reject pending step
      parameters:
        - in: path
          name: approvalId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision]
              properties:
                decision: { type: string, enum: [approved, rejected] }
                notes: { type: string, default: '' }
      responses:
        '200': { description: Decision applied, content: { application/json: { schema: { $ref: '#/components/schemas/OkResponse' } } } }
...
  /api/agentic/events:
    get:
      tags: [Agentic]
      summary: List task/tool events
      parameters:
        - in: query
          name: type
          schema: { type: string, enum: [task, tool], default: task }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 5000, default: 200 }
      responses:
        '200': { description: Events, content: { application/json: { schema: { $ref: '#/components/schemas/OkResponse' } } } }
...
  /api/agentic/mcp/servers:
    get:
      tags: [Agentic]
      summary: List MCP servers
      parameters:
        - in: query
          name: includeTools
          schema: { type: boolean, default: false }
      responses:
        '200': { description: MCP servers, content: { application/json: { schema: { $ref: '#/components/schemas/OkResponse' } } } }
  /api/agentic/mcp/servers/{serverId}/tools:
    get:
      tags: [Agentic]
      summary: List tools for an MCP server
      parameters:
        - in: path
          name: serverId
          required: true
          schema: { type: string }
      responses:
        '200': { description: MCP tools, content: { application/json: { schema: { $ref: '#/components/schemas/OkResponse' } } } }
  /api/agentic/mcp/servers/{serverId}/call:
    post:
      tags: [Agentic]
      summary: Execute an MCP tool
      parameters:
        - in: path
          name: serverId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tool]
              properties:
                tool: { type: string }
                input: { type: object }
      responses:
        '200': { description: MCP tool output, content: { application/json: { schema: { $ref: '#/components/schemas/OkResponse' } } } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    cookieAuth:
      type: apiKey
      in: cookie
      name: pd_session
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
  schemas:
    OkResponse:
      type: object
      properties:
        ok: { type: boolean, example: true }
        data:
          type: object
          additionalProperties: true
...
    AgenticTeam:
      type: object
      properties:
        mode: { type: string, enum: [auto, manual], default: auto }
        personaIds:
          type: array
          items: { type: string }
          default: []
        tags:
          type: array
          items: { type: string }
          default: []
        maxAgents: { type: integer, minimum: 1, maximum: 8, default: 3 }
    AgenticPlanRequest:
      type: object
      required: [goal]
      properties:
        goal: { type: string }
        constraints: { type: string, default: '' }
        maxSteps: { type: integer, minimum: 1, maximum: 12, default: 6 }
        team: { $ref: '#/components/schemas/AgenticTeam' }
    AgenticTaskStep:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        type: { type: string, enum: [tool, llm, job], default: tool }
        toolId: { type: string, default: '' }
        prompt: { type: string, default: '' }
        model: { type: string, default: '' }
        input: { type: object, additionalProperties: true, default: {} }
        requiresApproval: { type: boolean, default: false }
        dependsOn:
          type: array
          items: { type: string }
          default: []
    CreateAgenticTaskRequest:
      type: object
      required: [steps]
      properties:
        title: { type: string, default: Agentic Task }
        objective: { type: string, default: '' }
        team: { $ref: '#/components/schemas/AgenticTeam' }
        settings:
          type: object
          properties:
            model: { type: string, default: gpt-4.1-mini }
            temperature: { type: number, minimum: 0, maximum: 2, default: 0.3 }
        steps:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/AgenticTaskStep' }
        runImmediately: { type: boolean, default: false }
...